requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.1
python-Wappalyzer==0.3.1
tqdm==4.66.2
